+++ clients/python/inspr/__init__.py
@@ -1,4 +1,4 @@
 from .client import *
 from .rest import *
 from .models import *
-from .controller.controller_client import *
+from .controller.controller_client import ControllerClient