+++ clients/python/inspr/client.py
@@ -20,7 +20,7 @@ class Client:
             "data": msg
         }
         try:
-            send_post_request(self.write_address + "/" + channel, msg_body)
+            send_post_request(self.write_address + "/channel/" + channel, msg_body)
         except Exception as e:
             print(f"Error while trying to write message: {e}")
             raise Exception("failed to deliver message: channel: {}".format(channel))
@@ -37,7 +37,7 @@ class Client:
 
                 return '', HTTPStatus.OK
 
-            self.app.add_url_rule("/" + channel, endpoint = channel, view_func = route_func, methods=["POST"])
+            self.app.add_url_rule("/channel/" + channel, endpoint = channel, view_func = route_func, methods=["POST"])
             return handle_func
         return wrapper