+++ pkg/sidecars/lbsidecar/handlers.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"io/ioutil"
 	"net/http"
+	"net/url"
 	"os"
 	"strings"
 	"time"
@@ -93,6 +94,48 @@ func (s *Server) writeMessageHandler() rest.Handler {
 	}
 }
 
+// readMessageHandler handles requests sent to the read message server in the path
+// "/channel", for the lbsidecar
+func (s *Server) sendRequest() rest.Handler {
+	return func(w http.ResponseWriter, r *http.Request) {
+		path := strings.TrimPrefix(r.URL.Path, "/route/")
+		pathArgs := strings.Split(path, "/")
+		route := pathArgs[0]
+		endpoint := pathArgs[1]
+
+		logger.Info("handling route request on " + route)
+		resolved, err := environment.GetRouteData(route)
+
+		if err != nil {
+			logger.Error("unable to send request to "+route,
+				zap.Any("error", err))
+
+			rest.ERROR(w, err)
+		}
+
+		if !resolved.Endpoints.Contains(endpoint) {
+			err = ierrors.New("invalid endpoint: %s", endpoint).BadRequest()
+			logger.Error("unable to send request to "+path,
+				zap.Any("error", err))
+
+			rest.ERROR(w, err)
+		}
+		URL, _ := url.Parse(fmt.Sprintf("%s/%s", resolved.Address, path))
+		r.URL = URL
+		r.RequestURI = ""
+		r.Header.Set("X-Forwarded-For", r.RemoteAddr)
+		client := http.DefaultClient
+		resp, err := client.Do(r)
+		if err != nil {
+			rest.ERROR(w, err)
+			return
+		}
+		w.WriteHeader(resp.StatusCode)
+		io.Copy(w, resp.Body)
+		resp.Body.Close()
+	}
+}
+
 // readMessageHandler handles requests sent to the read message server
 func (s *Server) readMessageHandler() rest.Handler {
 	return func(w http.ResponseWriter, r *http.Request) {
@@ -151,6 +194,50 @@ func (s *Server) readMessageHandler() rest.Handler {
 		elapsed := time.Since(start)
 		s.GetMetric(channel).readMessageDuration.Observe(elapsed.Seconds())
 		s.GetMetric(channel).messagesRead.Add(1)
+	}
+}
+
+// routeReceiveHandler handles any requests received in the "/route" path, for the lbsidecar
+func (s *Server) routeReceiveHandler() rest.Handler {
+	return func(w http.ResponseWriter, r *http.Request) {
+		// Checking the endpoint
+		endpoint := strings.TrimPrefix(r.URL.Path, "/route/")
+
+		// port resolution: using the same as readHandler -> clientReadPort
+		clientReadPort := os.Getenv("INSPR_SCCLIENT_READ_PORT")
+		if clientReadPort == "" {
+			rest.ERROR(
+				w,
+				ierrors.New(
+					"[ENV VAR] INSPR_SCCLIENT_READ_PORT not found",
+				).NotFound(),
+			)
+			return
+		}
+
+		// Redirect the request
+		// localhost:port/route/endpoint
+		client := http.DefaultClient
+
+		URL, _ := url.Parse(fmt.Sprintf("http://localhost:%v/route/%v", clientReadPort, endpoint))
+		r.URL = URL
+		r.RequestURI = ""
+		r.Header.Set("X-Forwarded-For", r.RemoteAddr)
+
+		resp, err := client.Do(r)
+
+		// Validate the response
+		if err != nil {
+			logger.Error("route: unable to send request from lbsidecar to node",
+				zap.Any("error", err))
+			rest.ERROR(w, err)
+			return
+		}
+		defer resp.Body.Close()
+
+		// Return the response
+		w.WriteHeader(resp.StatusCode)
+		io.Copy(w, resp.Body)
 	}
 }