+++ pkg/sidecars/lbsidecar/server.go
@@ -153,12 +153,13 @@ func (s *Server) Run(ctx context.Context) error {
 		}
 	}()
 
-	mux := http.NewServeMux()
+	muxWriter := http.NewServeMux()
 
-	mux.Handle("/channel/", s.writeMessageHandler().Post().JSON())
+	muxWriter.Handle("/channel/", s.writeMessageHandler().Post().JSON())
+	muxWriter.Handle("/route/", s.sendRequest().JSON())
 
 	writeServer := &http.Server{
-		Handler: mux,
+		Handler: muxWriter,
 		Addr:    s.writeAddr,
 	}
 	go func() {
@@ -169,8 +170,13 @@ func (s *Server) Run(ctx context.Context) error {
 		}
 	}()
 
+	muxReader := http.NewServeMux()
+
+	muxReader.Handle("/channel/", s.readMessageHandler().Post().JSON())
+	muxReader.Handle("/route/", s.routeReceiveHandler().Post().JSON())
+
 	readServer := &http.Server{
-		Handler: s.readMessageHandler().Post().JSON(),
+		Handler: muxReader,
 		Addr:    s.readAddr,
 	}
 	go func() {